LPD-69683 Moving getLiferayTableNames method to DBResourceUtil as it can be reused for pre upgrade and post upgrade data cleanups
